Update sleep_detection.ino
Updated this part of my branch to essentially be able to first read 1 minute of data. I then signal process to remove any accidental extra peaks detected around the R part of the PQRS wave. Finally, I get the totals of each second and then divide by 60 to get the average heart rate that we can use as a metric to measure if we are below the heart rate for the states.